Separate check of empty yuzu_username and yuzu_token
yuzu_username is now automatically set to the current profile, so the previous check to see if yuzu_username is empty would leave the token as blank until you specifically hit the "Reset Token" button. Now with them separate, the default assignment of yuzu_username is just a backup since it will probably never be empty.

The token is separately checked, and a new one is generated if it's currently empty.